Search AGO web maps
This batch file executes the included Python script to help users search an ArcGIS Online organization for the use of specified layers and/or URLs. Double-click the batch file and enter the requested information.  The script will search users' content folders for web maps. It then scans the web maps layer list and basemap to find search terms entered by the user.